refactoring during cost-calculator package extraction
