+++ packages/modelfusion/src/model-provider/openai/OpenAIChatModel.ts
@@ -1,4 +1,3 @@
-import z from "zod";
 import {
   FlexibleObjectFromTextPromptTemplate,
   ObjectFromTextPromptTemplate,
@@ -25,7 +24,7 @@ import { countOpenAIChatPromptTokens } from "./countOpenAIChatMessageTokens";
 
 // https://platform.openai.com/docs/models
 // Open AI base chat models and their context window sizes.
-export const OPENAI_CHAT_CONTEXT_WINDOW_SIZES = {
+export const CHAT_MODEL_CONTEXT_WINDOW_SIZES = {
   "gpt-4": 8192,
   "gpt-4-0314": 8192,
   "gpt-4-0613": 8192,
@@ -45,137 +44,15 @@ export const OPENAI_CHAT_CONTEXT_WINDOW_SIZES = {
   "gpt-3.5-turbo-16k-0613": 16384,
 } as const;
 
-export type OpenAIChatModelBaseType =
-  keyof typeof OPENAI_CHAT_CONTEXT_WINDOW_SIZES;
-
-type FineTuneableOpenAIChatModelType =
-  | `gpt-3.5-turbo`
-  | `gpt-3.5-turbo-0613`
-  | `gpt-4-0613`;
-
-type FineTunedOpenAIChatModelType =
-  `ft:${FineTuneableOpenAIChatModelType}:${string}:${string}:${string}`;
-
-export type OpenAIChatBaseModelType =
-  keyof typeof OPENAI_CHAT_CONTEXT_WINDOW_SIZES;
-
-export interface OpenAIChatSettings extends AbstractOpenAIChatSettings {
-  model: OpenAIChatModelType;
-}
-
-export type OpenAIChatModelType =
-  | OpenAIChatBaseModelType
-  | FineTunedOpenAIChatModelType;
-
-export const openAIChatResponseSchema = z.object({
-  id: z.string(),
-  choices: z.array(
-    z.object({
-      message: z.object({
-        role: z.literal("assistant"),
-        content: z.string().nullable(),
-        function_call: z
-          .object({
-            name: z.string(),
-            arguments: z.string(),
-          })
-          .optional(),
-        tool_calls: z
-          .array(
-            z.object({
-              id: z.string(),
-              type: z.literal("function"),
-              function: z.object({
-                name: z.string(),
-                arguments: z.string(),
-              }),
-            })
-          )
-          .optional(),
-      }),
-      index: z.number().optional(), // optional for OpenAI compatible models
-      logprobs: z.nullable(z.any()),
-      finish_reason: z
-        .enum([
-          "stop",
-          "length",
-          "tool_calls",
-          "content_filter",
-          "function_call",
-        ])
-        .optional()
-        .nullable(),
-    })
-  ),
-  created: z.number(),
-  model: z.string(),
-  system_fingerprint: z.string().optional().nullable(),
-  object: z.literal("chat.completion"),
-  usage: z.object({
-    prompt_tokens: z.number(),
-    completion_tokens: z.number(),
-    total_tokens: z.number(),
-  }),
-});
-
-export type OpenAIChatResponse = z.infer<typeof openAIChatResponseSchema>;
-
-export const openaiChatChunkSchema = z.object({
-  object: z.string(), // generalized for openai compatible providers, z.literal("chat.completion.chunk")
-  id: z.string(),
-  choices: z.array(
-    z.object({
-      delta: z.object({
-        role: z.enum(["assistant", "user"]).optional(),
-        content: z.string().nullable().optional(),
-        function_call: z
-          .object({
-            name: z.string().optional(),
-            arguments: z.string().optional(),
-          })
-          .optional(),
-        tool_calls: z
-          .array(
-            z.object({
-              id: z.string(),
-              type: z.literal("function"),
-              function: z.object({
-                name: z.string(),
-                arguments: z.string(),
-              }),
-            })
-          )
-          .optional(),
-      }),
-      finish_reason: z
-        .enum([
-          "stop",
-          "length",
-          "tool_calls",
-          "content_filter",
-          "function_call",
-        ])
-        .nullable()
-        .optional(),
-      index: z.number(),
-    })
-  ),
-  created: z.number(),
-  model: z.string().optional(), // optional for OpenAI compatible models
-  system_fingerprint: z.string().optional().nullable(),
-});
-
-export type OpenAIChatChunk = z.infer<typeof openaiChatChunkSchema>;
-
 export function getOpenAIChatModelInformation(model: OpenAIChatModelType): {
   baseModel: OpenAIChatBaseModelType;
   isFineTuned: boolean;
   contextWindowSize: number;
 } {
   // Model is already a base model:
-  if (model in OPENAI_CHAT_CONTEXT_WINDOW_SIZES) {
+  if (model in CHAT_MODEL_CONTEXT_WINDOW_SIZES) {
     const contextWindowSize =
-      OPENAI_CHAT_CONTEXT_WINDOW_SIZES[model as OpenAIChatBaseModelType];
+      CHAT_MODEL_CONTEXT_WINDOW_SIZES[model as OpenAIChatBaseModelType];
 
     return {
       baseModel: model as OpenAIChatBaseModelType,
@@ -191,18 +68,37 @@ export function getOpenAIChatModelInformation(model: OpenAIChatModelType): {
   if (
     ["gpt-3.5-turbo", "gpt-3.5-turbo-0613", "gpt-4-0613"].includes(baseModel)
   ) {
-    const fineTunedBaseModel = baseModel as FineTuneableOpenAIChatModelType;
     const contextWindowSize =
-      OPENAI_CHAT_CONTEXT_WINDOW_SIZES[fineTunedBaseModel];
+      CHAT_MODEL_CONTEXT_WINDOW_SIZES[
+        baseModel as FineTuneableOpenAIChatModelType
+      ];
 
     return {
-      baseModel: fineTunedBaseModel,
+      baseModel: baseModel as FineTuneableOpenAIChatModelType,
       isFineTuned: true,
       contextWindowSize,
     };
   }
 
   throw new Error(`Unknown OpenAI chat base model ${baseModel}.`);
+}
+
+type FineTuneableOpenAIChatModelType =
+  | `gpt-3.5-turbo`
+  | `gpt-3.5-turbo-0613`
+  | `gpt-4-0613`;
+
+type FineTunedOpenAIChatModelType =
+  `ft:${FineTuneableOpenAIChatModelType}:${string}:${string}:${string}`;
+
+export type OpenAIChatBaseModelType =
+  keyof typeof CHAT_MODEL_CONTEXT_WINDOW_SIZES;
+
+export type OpenAIChatModelType =
+  | OpenAIChatBaseModelType
+  | FineTunedOpenAIChatModelType;
+export interface OpenAIChatSettings extends AbstractOpenAIChatSettings {
+  model: OpenAIChatModelType;
 }
 
 /**
